+++ tools/generate-classes/reservedEnums.js
@@ -1,5 +1,0 @@
-const reservedEnums = [
-  "OPAQUE", // OPAQUE is defined in wingdi.h
-];
-
-module.exports = reservedEnums;